Move auth/base helpers into base
